New comparison script
=== New
- New script to compare two results file in 'scripts' folder. It
compares the number of lines in the file, then for each line it compares
the number of elements (columns). If there is no difference, it compares
the values up to a tolerance (chosen in the beginning of the file). If
there is any difference in one of those comparisons (lines, columns,
values), the program exits with error 1, such that gitlab-CI recognizes
the error.