Added new test for time propagation, and now output folder is directly created
Improved running of tests in gitlab-ci (only one place to change).
Changed time-propagation test with smaller error (leading to smaller step size, so smaller integration time).
Changed many things in the code to try to get equal results with tolerance <1.d-8, but it was not possible.
Maybe it'll be better to change dimag back to aimag. TODO: Check which one is better for double complex to double real.
Removed -no-prec-sqrt from CMakeLists.

=== New
- TITAN now calls the command line to create the folder set up in the output variable
- New test for time propagation of Ni bulk
- Added option to input tolerance in comparison script. Used that to increase tolerance of time propagation

=== Changed
- gitlab-CI: Moved the information about compiler and python to before_script
- The tolerance of the comparison script was increased to 1.e-6, as some of the time-propagated results were bigger than that. Maybe it is a good idea to make the tolerance as a parameter for the script.
- Removed output folder from Nb_100_Monolayer example