Intermediate changes
There have been several modifications to the equations to calculate
Jij in q space. The advances so far lead to conclude that there was
a mistake in the second term used when we calculate Jij(q).

I am still not sure whether this modifications solve the problem
since now all the components of the DMI vanish for the system of a
magnetic monolayer on top of a Nb slab