Patch to the new J equations
As mentioned in the last commit. I was not so sure about the results
given by the new equations. There was a problem in the second term
coming from not considering the change in sign for the cell vector
R when the site indices are swaped. This was corrected in my notes
as well as in TITAN.

The results seem satisfactory and resemble the DMI given by KKR for
the same system of slab + Mn monolayer.

Now I need to check the equations for the superconducting term and
correct them accordingly.