Cleanup temporal comments
Throughout the development of the real space MEI calculations I had to
add several comments to print to terminal different variables or values.
Now that the development is ripe, it is time for those comments to go
away. In this comment I clean most of them. I have yet to add comments
related to the code and equations.